refactor: keep discord formatting local
